Improve docstrings in tests
