+++ tests/examples/cutting_stock_problem/model.py
@@ -1,4 +1,6 @@
 # pyright: reportAttributeAccessIssue=false
+"""Example Pyoframe formulation of the classic cutting stock problem."""
+
 import os
 from pathlib import Path
