+++ tests/examples/diet_problem/model.py
@@ -1,4 +1,4 @@
-"""Diet module example"""
+"""Example Pyoframe formulation of the classic diet problem."""
 
 # pyright: reportAttributeAccessIssue=false
