+++ tests/examples/portfolio_optim/model.py
@@ -1,5 +1,5 @@
 # pyright: reportAttributeAccessIssue=false
-"""Quadratic Portfolio Optimization Model (Markowitz Mean-Variance)
+"""Quadratic Portfolio Optimization Model (Markowitz Mean-Variance).
 
 Minimize portfolio variance subject to achieving a minimum return.
 
@@ -36,7 +36,7 @@ _input_dir = Path(os.path.dirname(os.path.realpath(__file__))) / "input_data"
 
 
 def solve_model(use_var_names=True, solver="gurobi"):
-    """Portfolio Optimization Test Case - Quadratic Programming with IPOPT
+    """Portfolio Optimization Test Case - Quadratic Programming with IPOPT.
 
     This example demonstrates IPOPT's capability to solve quadratic programming problems
     by implementing the classic Markowitz mean-variance portfolio optimization.
