+++ tests/util.py
@@ -1,3 +1,5 @@
+"""Utility functions for testing in Pyoframe."""
+
 from __future__ import annotations
 
 import ast
